Add autodetection of MAC address, NTP improvements, oui.txt cleanup
1) use ip link or ifconfig to select a MAC address if not user-specified
2) handle upper-case hex input
3) allow user to specify an NTP server
4) if ntpq fails, try fallback to ntpdate. (Current ntp servers often
   block 'rv'.commands.)
5) ntp utilities will try all IP addresses assigned to a host.
   Deal with multiple responses.
6) If oui.txt is downloaded, place it in /tmp and remove it on exit.
7) Fix use of undefined 'date' (instead of 'clock') in GID, use LSB of SHA per RFC
8) Use variables for defaults to simplify customization
9) Default to not prompting; use -i for prompts.
10) Add usage and command line options.
11) Ensure all errors reported to stderr.  Fully validate input.
12) Suppress leading zeros in ULA
13) Don't depend on GNU sed's extended REs or GNU echo -ne
14) By default, only output generated ULA.  -v for verbose mode.
15) Add option to get (download) a copy of oui.txt
Note that ntpq/ntpdate can take a while to get a timestamp.
ntpq may timeout, but the script will try to recover using ntpdate.

Manually providing a MAC address and/or timestamp should not be
necessary.  Doing so without understanding the RFC is likely to
reduce the uniqueness of the generated address.  -i is provided
for advanced users.

The MAC autoselection uses the first interface with an ethernet
address. The address of another interface can be used safely.
However, since the purpose is only to seed a hash to identify
the network, there is no reason to prefer one over another.

The timestamp ensures that a given interface can generate more than
one ULA.  The only reason to set it manually is deterministic
debugging
.